Fix React invariant exception
Solved this by using {reactive: false} in the Mongo query we use
to fetch the rectangle data, and then forcing a re-render of the
rectangle parent component with a new key.

This means that if other people update the rectangle you'll need
a refresh at the moment to see new photos... We'll have to fix that?